Refactor/reorg.
